#66 refactor(domain/functions): utils의 순수 함수 3종을 domain/functions로 이관
utils/ 의 순수 함수들을 도메인 레이어로 격상시켜 store/api 의존 없는 명시적 경계를 만든다.

- src/utils/repeatingTimeCalculator.ts → src/domain/functions/repeating.ts
  (prefer-const 린트 오류 수정 포함)
- src/utils/eventTimeUtils.ts         → src/domain/functions/eventTime.ts
- src/utils/tagDisplay.ts             → src/domain/functions/tagDisplay.ts
  TFunction 의존 제거 — i18next 타입 대신 로컬 TranslateFn 인터페이스 사용
- 관련 테스트 파일 tests/utils/ → tests/domain/functions/ 이동 후 import 경로 갱신

이관된 파일들은 store/api/i18n/repositories 를 직접 import하지 않음.
외부 사용처(components, calendar, repositories, stores, utils)의 import 경로 전수 갱신.
