+++ src/calendar/MainCalendarGrid.tsx
@@ -1,8 +1,8 @@
 import { useMemo, useRef, useState, useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import type { CalendarDay } from './calendarUtils'
-import type { CalendarEvent } from '../utils/eventTimeUtils'
-import { formatDateKey } from '../utils/eventTimeUtils'
+import type { CalendarEvent } from '../domain/functions/eventTime'
+import { formatDateKey } from '../domain/functions/eventTime'
 import { buildWeekEventStack, type EventOnWeekRow } from './weekEventStackBuilder'
 import { useUiStore } from '../stores/uiStore'
 import { useCalendarEventsCache } from '../repositories/caches/calendarEventsCache'
