+++ src/components/DayEventList.tsx
@@ -4,14 +4,14 @@ import { todoApi } from '../api/todoApi'
 import { useCalendarEventsCache } from '../repositories/caches/calendarEventsCache'
 import { useTagFilterStore } from '../stores/tagFilterStore'
 import { useResolvedEventTag } from '../hooks/useResolvedEventTag'
-import { tagDisplayName } from '../utils/tagDisplay'
+import { tagDisplayName } from '../domain/functions/tagDisplay'
 import { TimeDescription } from './TimeDescription'
 import { RepeatingScopeDialog, type RepeatScope } from './RepeatingScopeDialog'
-import { nextRepeatingTime, getStartTimestamp } from '../utils/repeatingTimeCalculator'
+import { nextRepeatingTime, getStartTimestamp } from '../domain/functions/repeating'
 import { refreshTodoListStores } from '../utils/todoActions'
-import { formatDateKey } from '../utils/eventTimeUtils'
+import { formatDateKey } from '../domain/functions/eventTime'
 import { useSettingsCache } from '../repositories/caches/settingsCache'
-import type { CalendarEvent } from '../utils/eventTimeUtils'
+import type { CalendarEvent } from '../domain/functions/eventTime'
 import type { Todo } from '../models'
 
 function EventItem({ calEvent, onEventClick, onComplete, isLast }: {
