+++ src/components/EventDetailPopover.tsx
@@ -3,10 +3,10 @@ import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import type { TFunction } from 'i18next'
 import { Pencil, Trash2, X, Clock, Repeat, Bell, MapPin, FileText, Link2 } from 'lucide-react'
-import type { CalendarEvent } from '../utils/eventTimeUtils'
+import type { CalendarEvent } from '../domain/functions/eventTime'
 import type { Repeating, NotificationOption } from '../models'
 import { useResolvedEventTag } from '../hooks/useResolvedEventTag'
-import { tagDisplayName } from '../utils/tagDisplay'
+import { tagDisplayName } from '../domain/functions/tagDisplay'
 import { EventTimeDisplay } from './EventTimeDisplay'
 import { eventDetailApi } from '../api/eventDetailApi'
 import type { EventDetail } from '../models'
