+++ src/components/RightEventPanel.tsx
@@ -1,8 +1,8 @@
 import { useTranslation } from 'react-i18next'
 import { CheckCircle2 } from 'lucide-react'
 import { useUiStore } from '../stores/uiStore'
-import { formatDateKey } from '../utils/eventTimeUtils'
-import type { CalendarEvent } from '../utils/eventTimeUtils'
+import { formatDateKey } from '../domain/functions/eventTime'
+import type { CalendarEvent } from '../domain/functions/eventTime'
 import { useForemostEventCache } from '../repositories/caches/foremostEventCache'
 import { useHolidayCache } from '../repositories/caches/holidayCache'
 import { useSettingsCache } from '../repositories/caches/settingsCache'
