+++ src/pages/TodoFormPage.tsx
@@ -10,7 +10,7 @@ import { useToastStore } from '../stores/toastStore'
 import { deleteTodoEvent } from '../utils/eventDeleteHelper'
 import { ConfirmDialog } from '../components/ConfirmDialog'
 import { RepeatingScopeDialog, type RepeatScope } from '../components/RepeatingScopeDialog'
-import { nextRepeatingTime, getStartTimestamp } from '../utils/repeatingTimeCalculator'
+import { nextRepeatingTime, getStartTimestamp } from '../domain/functions/repeating'
 import { EventFormHeader } from '../components/eventForm/EventFormHeader'
 import { EventTimeSection } from '../components/eventForm/EventTimeSection'
 import { EventDetailsSection } from '../components/eventForm/EventDetailsSection'
